Fix make run-devel: Refactor application structure by moving the main FastAPI app logic to main.py and updating imports accordingly
