+++ Makefile
@@ -16,12 +16,14 @@ pre-init:
 	cd client && npm install
 	mkdir -p server/movici_viewer/ui
 
+
+# 	TODO: resolve inconsistency on min python version between pyproject.toml and poetry env use. 
+# 	&& poetry env use python3.13 \ 
 init: pre-init ui
 	cd server \
-	&& poetry env use python3.10 \
 	&& poetry install --extras dev
 
-data_dir=tests/data
+data_dir=/tests/data
 port=5000
 
 export data_dir
@@ -36,5 +38,6 @@ run-devel:
 run-client:
 	cd client \
 	&& VITE_MOVICI_BASE_URL=http://localhost:$(port) npm run dev
+	
 run:
 	cd server && poetry run movici-viewer
